refactor(frontend): split wasm app module and keep native app testable
